Codebase maintainability improvements (tasks 1.1-2.4) + issue #15
Completed improvements:
- Task 1.1: Reduce unsafe unwrap/expect calls across 14 files
- Task 1.2: Comprehensive error types for 4 crates (thiserror)
- Task 1.3: Null/invalid state guards (debug_assert, NaN/Inf checks)
- Task 2.1: Address all clippy warnings (155 → 0)
- Task 2.2: Split large files (20 → 13 files >1000 LOC)
- Task 2.3: Eliminate temporary debug code (println/eprintln)
- Task 2.4: Complete TODO/FIXME items (15 → 9 remaining)

Issue #15 - Editor error dialogs:
- Created shared show_error_dialog() in file_dialog.rs
- User-friendly messages with file paths, modal + transient_for
- Consolidated 5 duplicate implementations across UI modules

Additional:
- FileChooserNative → FileChooserDialog (KDE/Kubuntu fix)
- Dirty flag for 3D visualizer buffer updates (#18)
- Firmware settings JSON file load/save (#13)

+++ crates/gcodekit5-designer/src/designer_state/file_io.rs
@@ -65,7 +65,7 @@ impl DesignerState {
 
         // Restore shapes with full DrawingObject (preserves all properties)
         for shape_data in &design.shapes {
-            if let Ok(obj) = DesignFile::to_drawing_object(shape_data, shape_data.id as i32) {
+            if let Ok(obj) = DesignFile::to_drawing_object(shape_data, shape_data.id) {
                 self.canvas.restore_shape(obj);
             }
         }
